Refactor project structure
Move cli cmd_*.go files to `package cli`.
Keep entry point at `cmd/falcon/main.go`.
