+++ cli/info.go
