+++ cli/info_test.go
