+++ cli/sign.go
