+++ cli/sign_test.go
