+++ cli/verify.go
