+++ cli/verify_test.go
